requirements.txt
# Discord library with application commands (slash commands) support
discord.py>=2.3.2

# Async SQLite driver for non-blocking database operations
aiosqlite>=0.19.0

# Image generation for the heatmap grid
Pillow>=10.0.0

# Timezone support for /set_timezone and local-time views
pytz>=2024.1

# Load bot token from .env file
python-dotenv>=1.0.0
